YY - yay. "hot mass" working, control working, simulation its own moudle in game
